Move common menu templates to public API
