+++ apps/files_sharing/lib/Controller/ShareController.php
@@ -38,9 +38,9 @@ namespace OCA\Files_Sharing\Controller;
 use OC_Files;
 use OC_Util;
 use OCA\FederatedFileSharing\FederatedShareProvider;
-use OCA\Files_Sharing\Template\ExternalShareMenuAction;
-use OCA\Files_Sharing\Template\LinkMenuAction;
 use OCP\AppFramework\Http\Template\SimpleMenuAction;
+use OCP\AppFramework\Http\Template\ExternalShareMenuAction;
+use OCP\AppFramework\Http\Template\LinkMenuAction;
 use OCP\AppFramework\Http\Template\PublicTemplateResponse;
 use OCP\Defaults;
 use OCP\IL10N;
